inkscape_extension/requirements.txt
lxml>=4.9.0
pyyaml>=6.0
numpy>=1.22.0 